# docs/swagger.yaml
basePath: /
definitions:
...
  controller.AdminAuditLogListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/controller.AdminAuditLogResponse'
        type: array
      next_cursor:
        $ref: '#/definitions/controller.AdminAuditLogCursorDTO'
    type: object
...
  controller.AdminEventResponse:
    properties:
      created_at:
        type: string
      end_at:
        type: string
      id:
        type: integer
      name:
        type: string
      organizer_id:
        type: integer
      sale_end_at:
        type: string
      sale_start_at:
        type: string
      start_at:
        type: string
      status:
        type: integer
      updated_at:
        type: string
      venue:
        type: string
    type: object
...
  controller.AdminOrderListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/controller.AdminOrderResponse'
        type: array
      next_cursor:
        $ref: '#/definitions/controller.AdminOrderListCursorDTO'
    type: object
...
  controller.AdminSectionResponse:
    properties:
      available_quantity:
        type: integer
      created_at:
        type: string
      event_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
      purchase_limit:
        type: integer
      reserved_quantity:
        type: integer
      sold_quantity:
        type: integer
      status:
        type: integer
      total_quantity:
        type: integer
      updated_at:
        type: string
    type: object
...
  controller.CreateAdminEventRequest:
    properties:
      end_at:
        type: string
      name:
        type: string
      organizer_id:
        type: integer
      sale_end_at:
        type: string
      sale_start_at:
        type: string
      start_at:
        type: string
      status:
        type: integer
      venue:
        type: string
    type: object
  controller.CreateAdminSectionRequest:
    properties:
      name:
        type: string
      price:
        type: integer
      purchase_limit:
        type: integer
      status:
        type: integer
      total_quantity:
        type: integer
    type: object
...
  controller.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
      request_id:
        type: string
    type: object
...
  controller.RevealSensitiveRequest:
    properties:
      reason:
        type: string
    type: object
...
info:
  contact: {}
  description: 搶票系統 API
  title: buy-ticket API
  version: "1.0"
paths:
  /admin/audit-logs:
    get:
      description: List admin audit logs with keyset pagination. SUPER_ADMIN can see
        all logs; EVENT_ADMIN can only see its own logs.
      parameters:
      - description: Page size, default 20, max 100
        in: query
        name: limit
        type: integer
      - description: Cursor created_at in RFC3339
        in: query
        name: cursor_created_at
        type: string
      - description: Cursor audit log id
        in: query
        name: cursor_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.AdminAuditLogListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: List admin audit logs
      tags:
      - admin-audit-logs
...
  /admin/events:
    get:
      description: List events for admin backoffice. EVENT_ADMIN is scoped to its
        organizer.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controller.AdminEventResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: List admin events
      tags:
      - admin-events
    post:
      consumes:
      - application/json
      description: Create event for admin backoffice. SUPER_ADMIN must provide organizer_id;
        EVENT_ADMIN uses its own organizer.
      parameters:
      - description: Event payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.CreateAdminEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.AdminEventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Create admin event
      tags:
      - admin-events
  /admin/events/{eventId}:
    get:
      description: Get an event for admin backoffice. EVENT_ADMIN is scoped to its
        organizer.
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.AdminEventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Get admin event
      tags:
      - admin-events
  /admin/events/{eventId}/sections:
    get:
      description: List event sections with inventory quantities. EVENT_ADMIN is scoped
        to its organizer.
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controller.AdminSectionResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: List admin event sections
      tags:
      - admin-events
    post:
      consumes:
      - application/json
      description: Create section for an event. EVENT_ADMIN is scoped to its organizer.
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: integer
      - description: Section payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.CreateAdminSectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.AdminSectionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Create admin event section
      tags:
      - admin-events
...
  /admin/orders:
    get:
      description: List orders for admin backoffice with keyset pagination. EVENT_ADMIN
        is scoped to its organizer.
      parameters:
      - description: Page size, default 20, max 100
        in: query
        name: limit
        type: integer
      - description: Cursor created_at in RFC3339
        in: query
        name: cursor_created_at
        type: string
      - description: Cursor order id
        in: query
        name: cursor_id
        type: integer
      - description: Filter by event id
        in: query
        name: event_id
        type: integer
      - description: Filter by user id
        in: query
        name: user_id
        type: integer
      - description: Filter by order status
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.AdminOrderListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: List admin orders
      tags:
      - admin-orders
  /admin/orders/{orderId}/reveal-sensitive:
    post:
      consumes:
      - application/json
      description: Reveal full user information for an order. Only SUPER_ADMIN can
        use this endpoint and a reason is required.
      parameters:
      - description: Order ID
        in: path
        name: orderId
        required: true
        type: integer
      - description: Reveal reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.RevealSensitiveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.AdminOrderSensitiveResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Reveal order sensitive data
      tags:
      - admin-orders
  /admin/organizers:
    get:
      description: List organizers. Only SUPER_ADMIN can use this endpoint.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controller.OrganizerResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: List organizers
      tags:
      - admin-organizers
    post:
      consumes:
      - application/json
      description: Create an organizer. Only SUPER_ADMIN can use this endpoint.
      parameters:
      - description: Organizer payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.CreateOrganizerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.OrganizerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Create organizer
      tags:
      - admin-organizers
  /admin/users:
    get:
      description: List backoffice admin users. Only SUPER_ADMIN can use this endpoint.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controller.AdminUserResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: List admin users
      tags:
      - admin-users
    post:
      consumes:
      - application/json
      description: Create a backoffice admin user. Only SUPER_ADMIN can use this endpoint.
      parameters:
      - description: Admin user payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.CreateAdminUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.AdminUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Create admin user
      tags:
      - admin-users
  /events:
    get:
      description: 取得目前系統中的活動列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controller.EventResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 查詢活動列表
      tags:
      - events
  /events/{eventId}:
    get:
      description: 依活動 ID 取得活動基本資訊
      parameters:
      - description: 活動 ID
        in: path
        name: eventId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.EventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 取得活動資訊
      tags:
      - events
...
  /events/{eventId}/sections:
    get:
      description: 依活動 ID 取得票區列表
      parameters:
      - description: 活動 ID
        in: path
        name: eventId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controller.SectionResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 取得活動票區
      tags:
      - events
...
  /orders:
    post:
      consumes:
      - application/json
      description: 由 reservation 建立待付款訂單
      parameters:
      - description: 建立訂單請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.OrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 建立訂單
      tags:
      - orders
...
swagger: "2.0"